warlock: fix and improve
Display Burning Embers on the Fire and Brimstone versions of
Conflagrate, Immolate and Incinerate.
Don't create double rules for Burning Embers on Chaos Bolt.
Pyroclasm got merged into Backdraft for WoD.
Display the duration of Backdraft on Chaos Bolt when at or abvoe 3
stacks too.
Remove the MoltenCore rule (this is handled by the default ui).
Stop suggesting using Havoc when it is off cooldown. This is useless on
single target fights.